@@ -2,12 +2,12 @@
 name = panoptes-pocs
 description = PANOPTES Observatory Control System
 author = Project PANOPTES
-author-email = developers@projectpanoptes.org
+author_email = developers@projectpanoptes.org
 license = mit
-long-description = file: README.md
-long-description-content-type = text/markdown; charset=UTF-8
+long_description = file: README.md
+long_description_content_type = text/markdown; charset=UTF-8
 url = https://github.com/panoptes/POCS
-project-urls =
+project_urls =
     Documentation = https://panoptes-pocs.readthedocs.org
     Forum = https://forum.projectpanoptes.org
 platforms = linux
@@ -18,22 +18,18 @@ classifiers =
     License :: OSI Approved :: MIT License
     Operating System :: POSIX
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3 :: Only
     Topic :: Scientific/Engineering :: Astronomy
     Topic :: Scientific/Engineering :: Physics
 
 [options]
 zip_safe = False
-packages = find:
 include_package_data = True
 package_dir =
     =src
 scripts =
-    bin/panoptes-develop
-    bin/pocs
-    bin/pocs-shell
-    bin/peas-shell
+    scripts/take-pic.sh
 
 # DON'T CHANGE THE FOLLOWING LINE! IT WILL BE UPDATED BY PYSCAFFOLD!
 
@@ -42,13 +38,18 @@ setup_requires = pyscaffold>=3.2a0,<3.3a0
 install_requires =
     astroplan
     astropy
-    panoptes-utils>=0.2.30
+    fastapi
+    panoptes-utils[config]>=0.2.35
+    pandas
+    Pillow>=8.3.2
     pyserial
+    requests
     transitions
+    uvicorn[standard]
 # The usage of test_requires is discouraged, see `Dependency Management` docs
 # tests_require = pytest; pytest-cov
 # Require a specific Python version, e.g. Python 2.7 or >= 3.4
-python_requires = >=3.7
+python_requires = >=3.9
 
 [options.packages.find]
 where = src
@@ -73,26 +74,17 @@ testing =
     pytest-doctestplus
     pytest-remotedata>=0.3.1
     responses
+sensors =
+    streamz
 
 [options.entry_points]
-# Add here console scripts like:
-# console_scripts =
-#     script_name = panoptes.pocs.module:function
-# For example:
-# console_scripts =
-#     fibonacci = panoptes.pocs.skeleton:run
-# And any other entry points, for example:
-# pyscaffold.cli =
-#     awesome = pyscaffoldext.awesome.extension:AwesomeExtension
-
-[test]
-# py.test options when running `python setup.py test`
-extras = True
+console_scripts =
+    pocs = panoptes.pocs.utils.cli.main:app
 
 [tool:pytest]
 addopts =
     --cov panoptes.pocs
-    --cov panoptes.peas
+    --cov tests
     --cov-branch
     --cov-report term-missing:skip-covered
     --cov-report xml:build/coverage.xml
