@@ -8,12 +8,10 @@ import numpy as np
 from scipy.ndimage import binary_dilation
 from astropy.modeling import models
 from astropy.modeling import fitting
-
 from panoptes.pocs.base import PanBase
 from panoptes.utils.time import current_time
 from panoptes.utils.images import focus as focus_utils
 from panoptes.utils.images import mask_saturated
-
 from panoptes.pocs.utils.plotting import make_autofocus_plot
 
 
@@ -170,11 +168,25 @@ class AbstractFocuser(PanBase, metaclass=ABCMeta):
 
     @abstractmethod
     def move_to(self, position):
-        """ Move focuser to new encoder position """
+        """ Move focuser to new encoder position.
+
+        Args:
+            position (int): new focuser position, in encoder units.
+
+        Returns:
+            int: focuser position following the move, in encoder units.
+        """
         raise NotImplementedError
 
     def move_by(self, increment):
-        """ Move focuser by a given amount """
+        """ Move focuser by a given amount.
+
+        Args:
+            increment (int): distance to move the focuser, in encoder units.
+
+        Returns:
+            int: focuser position following the move, in encoder units.
+        """
         return self.move_to(self.position + increment)
 
     def autofocus(self,
@@ -189,6 +201,7 @@ class AbstractFocuser(PanBase, metaclass=ABCMeta):
                   mask_dilations=None,
                   coarse=False,
                   make_plots=None,
+                  filter_name=None,
                   blocking=False):
         """
         Focuses the camera using the specified merit function. Optionally performs
@@ -221,6 +234,8 @@ class AbstractFocuser(PanBase, metaclass=ABCMeta):
             make_plots (bool, optional): Whether to write focus plots to images folder. If not
                 given will fall back on value of `autofocus_make_plots` set on initialisation,
                 and if it wasn't set then will default to False.
+            filter_name (str, optional): The filter to use for focusing. If not provided, will use
+                last light position.
             blocking (bool, optional): Whether to block until autofocus complete, default False.
 
         Returns:
@@ -230,7 +245,8 @@ class AbstractFocuser(PanBase, metaclass=ABCMeta):
             ValueError: If invalid values are passed for any of the focus parameters.
         """
         self.logger.debug('Starting autofocus')
-        assert self._camera.is_connected, self.logger.error("Camera must be connected for autofocus!")
+        assert self._camera.is_connected, self.logger.error(
+            "Camera must be connected for autofocus!")
 
         assert self.is_connected, self.logger.error("Focuser must be connected for autofocus!")
 
@@ -294,6 +310,23 @@ class AbstractFocuser(PanBase, metaclass=ABCMeta):
 
         if make_plots is None:
             make_plots = self.autofocus_make_plots
+
+        # Move filterwheel to the correct position
+        if self.camera is not None:
+            if self.camera.has_filterwheel:
+
+                if filter_name is None:
+                    # NOTE: The camera will move the FW to the last light position automatically
+                    self.logger.warning(f"Filter name not provided for autofocus on {self}. Using"
+                                        " last light position.")
+                else:
+                    self.logger.info(f"Moving filterwheel to {filter_name} for autofocusing on"
+                                     f" {self}.")
+                    self.camera.filterwheel.move_to(filter_name, blocking=True)
+
+            elif filter_name is None:
+                self.logger.warning(f"Filter {filter_name} requiested for autofocus but"
+                                    f" {self.camera} has no filterwheel.")
 
         # Set up the focus parameters
         focus_event = Event()
@@ -402,12 +435,12 @@ class AbstractFocuser(PanBase, metaclass=ABCMeta):
         # Get focus steps.
         focus_positions = np.arange(max(initial_focus - focus_range / 2, self.min_position),
                                     min(initial_focus + focus_range / 2, self.max_position) + 1,
-                                    focus_step, dtype=np.int)
+                                    focus_step, dtype=int)
         n_positions = len(focus_positions)
 
         # Set up empty array holders
         cutouts = np.zeros((n_positions, cutout_size, cutout_size), dtype=initial_cutout.dtype)
-        masks = np.empty((n_positions, cutout_size, cutout_size), dtype=np.bool)
+        masks = np.empty((n_positions, cutout_size, cutout_size), dtype=bool)
         metrics = np.empty(n_positions)
 
         # Take and store an exposure for each focus position.
