@@ -6,15 +6,13 @@ from datetime import datetime
 from astropy import units as u
 from astropy.coordinates import get_moon
 from astropy.coordinates import get_sun
-
 from panoptes.pocs.base import PanBase
 from panoptes.pocs.camera import AbstractCamera
 from panoptes.pocs.dome import AbstractDome
 from panoptes.pocs.images import Image
-from panoptes.pocs.mount import AbstractMount
+from panoptes.pocs.mount.mount import AbstractMount
 from panoptes.pocs.scheduler import BaseScheduler
 from panoptes.pocs.utils.location import create_location_from_config
-
 from panoptes.utils.time import current_time
 from panoptes.utils import error
 
@@ -342,14 +340,14 @@ class Observatory(PanBase):
             return None
 
         # If observation list is empty or a reread is requested
-        reread_fields_file = (
+        reread_file = (
                 self.scheduler.has_valid_observations is False or
-                kwargs.get('reread_fields_file', False) or
+                kwargs.get('read_file', False) or
                 self.get_config('scheduler.check_file', default=False)
         )
 
         # This will set the `current_observation`
-        self.scheduler.get_observation(reread_fields_file=reread_fields_file, *args, **kwargs)
+        self.scheduler.get_observation(read_file=reread_file, *args, **kwargs)
 
         if self.current_observation is None:
             self.scheduler.clear_available_observations()
@@ -646,34 +644,29 @@ class Observatory(PanBase):
             cameras = {cam_name: self.cameras[
                 cam_name] for cam_name in camera_list if cam_name in self.cameras.keys()}
             if cameras == {}:
-                self.logger.warning(
-                    "Passed a list of camera names ({}) but no matches found".format(camera_list))
+                self.logger.warning(f"No matching camera names in ({camera_list})")
         else:
-            # No cameras specified, will try to autofocus all cameras from
-            # self.cameras
+            # No cameras specified, will try to autofocus all cameras from self.cameras
             cameras = self.cameras
 
         autofocus_events = dict()
 
         # Start autofocus with each camera
         for cam_name, camera in cameras.items():
-            self.logger.debug("Autofocusing camera: {}".format(cam_name))
+            self.logger.debug(f"Autofocusing camera: {cam_name}")
 
             try:
                 assert camera.focuser.is_connected
             except AttributeError:
-                self.logger.debug(
-                    'Camera {} has no focuser, skipping autofocus'.format(cam_name))
+                self.logger.debug(f'Camera {cam_name} has no focuser, skipping autofocus')
             except AssertionError:
-                self.logger.debug(
-                    'Camera {} focuser not connected, skipping autofocus'.format(cam_name))
+                self.logger.debug(f'Camera {cam_name} focuser not connected, skipping autofocus')
             else:
                 try:
                     # Start the autofocus
                     autofocus_event = camera.autofocus(**kwargs)
                 except Exception as e:
-                    self.logger.error(
-                        "Problem running autofocus: {}".format(e))
+                    self.logger.error(f"Problem running autofocus: {e!r}")
                 else:
                     autofocus_events[cam_name] = autofocus_event
 
