@@ -3,7 +3,7 @@ from pkg_resources import get_distribution, DistributionNotFound
 
 try:
     # Change here if project is renamed and does not equal the package name
-    dist_name = 'POCS'
+    dist_name = 'panoptes-pocs'
     __version__ = get_distribution(dist_name).version
 except DistributionNotFound:
     __version__ = 'unknown'
