@@ -1,10 +1,11 @@
-import copy
 from collections import OrderedDict
 import re
 import shutil
 import subprocess
 import random
 from contextlib import suppress
+
+import requests
 
 from panoptes.pocs.camera.camera import AbstractCamera  # noqa
 
@@ -16,7 +17,7 @@ from panoptes.utils.library import load_module
 logger = get_logger()
 
 
-def list_connected_cameras():
+def list_connected_cameras(remote_url=None):
     """Detect connected cameras.
 
     Uses gphoto2 to try and detect which cameras are connected. Cameras should
@@ -26,12 +27,17 @@ def list_connected_cameras():
         list: A list of the ports with detected cameras.
     """
 
-    gphoto2 = shutil.which('gphoto2')
-    if not gphoto2:  # pragma: no cover
-        raise error.NotFound('The gphoto2 command is missing, please install.')
-    command = [gphoto2, '--auto-detect']
-    result = subprocess.check_output(command)
-    lines = result.decode('utf-8').split('\n')
+    if remote_url is not None:
+        response = requests.post(remote_url, json=dict(arguments='--auto-detect'))
+        if response.ok:
+            result = response.json()['output']
+    else:
+        gphoto2 = shutil.which('gphoto2')
+        if not gphoto2:  # pragma: no cover
+            raise error.NotFound('gphoto2 is missing, please install or use the remote_url option.')
+        command = [gphoto2, '--auto-detect']
+        result = subprocess.check_output(command).decode('utf-8')
+    lines = result.split('\n')
 
     ports = []
 
@@ -100,12 +106,13 @@ def create_cameras_from_config(config=None,
     if auto_detect:
         logger.debug("Auto-detecting ports for cameras")
         try:
-            ports = list_connected_cameras()
+            ports = list_connected_cameras(remote_url=camera_defaults.get('remote_url', None))
         except error.PanError as e:
             logger.warning(e)
 
         if len(ports) == 0:
-            raise error.CameraNotFound(msg="No cameras detected. For testing, use camera simulator.")
+            raise error.CameraNotFound(
+                msg="No cameras detected. For testing, use camera simulator.")
         else:
             logger.debug(f"Detected ports={ports!r}")
 
@@ -151,9 +158,9 @@ def create_cameras_from_config(config=None,
                 else:
                     raise error.NotFound(f'module={module!r} does not have a Camera object')
         except error.NotFound:
-            logger.error(f"Cannot find camera module with config: {device_config}")
+            logger.error(f'Cannot find camera module with config: {device_config}')
         except Exception as e:
-            logger.error(f"Cannot create camera type: {model} {e}")
+            logger.error(f'Cannot create camera type: {model} {e!r}')
         else:
             # Check if the config specified a primary camera and if it matches.
             if camera.uid == camera_config.get('primary'):
@@ -169,7 +176,7 @@ def create_cameras_from_config(config=None,
 
     # If no camera was specified as primary use the first
     if primary_camera is None and auto_primary:
-        logger.info(f'No primary camera given, assigning the first camera (auto_primary={auto_primary!r})')
+        logger.info(f'No primary camera given, assigning the first camera ({auto_primary!r})')
         primary_camera = list(cameras.values())[0]  # First camera
         primary_camera.is_primary = True
 
