@@ -1,42 +1,44 @@
-import os
-import subprocess
-from abc import ABC
-from threading import Event
-from threading import Timer
-
 from astropy import units as u
-from panoptes.pocs.camera.gphoto import AbstractGPhotoCamera
+from panoptes.pocs.camera.gphoto.base import AbstractGPhotoCamera
 from panoptes.utils import error
-from panoptes.utils.images import cr2 as cr2_utils
-from panoptes.utils.time import CountdownTimer
 from panoptes.utils.time import current_time
 from panoptes.utils.utils import get_quantity_value
 
 
-class Camera(AbstractGPhotoCamera, ABC):
+class Camera(AbstractGPhotoCamera):
 
-    def __init__(self, *args, **kwargs):
-        kwargs['readout_time'] = 6.0
-        kwargs['file_extension'] = 'cr2'
+    def __init__(self, readout_time: float = 1.0, file_extension: str = 'cr2', connect: bool = True,
+                 *args, **kwargs):
+        """Create a camera object for a Canon EOS DSLR.
+
+        Args:
+            readout (float): The time it takes to read out the file from the
+                camera, default 1.0 second.
+            file_extension (str): The file extension to use, default `cr2`.
+            connect (bool): Connect to camera on startup, default True.
+        """
+        kwargs['readout_time'] = readout_time
+        kwargs['file_extension'] = file_extension
         super().__init__(*args, **kwargs)
+        self.logger.debug("Creating Canon DSLR GPhoto2 camera")
 
-        # Hold on to the exposure process for polling.
-        self._exposure_proc = None
-
-        self.logger.debug("Connecting GPhoto2 camera")
-        self.connect()
-        self.logger.debug(f"{self.name} connected")
+        if connect:
+            self.connect()
 
     @property
     def bit_depth(self):
         return 12 * u.bit
 
-    def connect(self):
-        """Connect to Canon DSLR
+    @property
+    def egain(self):
+        return 1.5 * (u.electron / u.adu)
 
-        Gets the serial number from the camera and sets various settings
+    def connect(self):
+        """Connect to Canon DSLR.
+
+        Gets the serial number from the camera and sets various settings.
         """
-        self.logger.debug('Connecting to camera')
+        self.logger.debug('Connecting to Canon gphoto2 camera')
 
         # Get serial number
         _serial_number = self.get_property('serialnumber')
@@ -47,153 +49,72 @@ class Camera(AbstractGPhotoCamera, ABC):
 
         # Properties to be set upon init.
         prop2index = {
-            '/main/actions/viewfinder': 1,  # Screen off
             '/main/capturesettings/autoexposuremode': 3,  # 3 - Manual; 4 - Bulb
             '/main/capturesettings/continuousaf': 0,  # No auto-focus
             '/main/capturesettings/drivemode': 0,  # Single exposure
             '/main/capturesettings/focusmode': 0,  # Manual (don't try to focus)
-            '/main/capturesettings/shutterspeed': 0,  # Bulb
             '/main/imgsettings/imageformat': 9,  # RAW
             '/main/imgsettings/imageformatcf': 9,  # RAW
             '/main/imgsettings/imageformatsd': 9,  # RAW
-            '/main/imgsettings/iso': 1,  # ISO 100
-            '/main/settings/autopoweroff': 0,  # Don't power off
             '/main/settings/capturetarget': 0,  # Capture to RAM, for download
-            '/main/settings/datetime': 'now',  # Current datetime
-            '/main/settings/datetimeutc': 'now',  # Current datetime
             '/main/settings/reviewtime': 0,  # Screen off after taking pictures
+            '/main/imgsettings/iso': 1,  # ISO 100
+            '/main/capturesettings/shutterspeed': 0,  # Bulb
         }
 
         owner_name = 'Project PANOPTES'
         artist_name = self.get_config('pan_id', default=owner_name)
-        copyright = f'{owner_name} {current_time().datetime:%Y}'
+        copy_right = f'{owner_name} {current_time().datetime:%Y}'
 
         prop2value = {
             '/main/settings/artist': artist_name,
-            '/main/settings/copyright': copyright,
+            '/main/settings/copyright': copy_right,
             '/main/settings/ownername': owner_name,
+            '/main/actions/syncdatetime': "1",
         }
 
-        self.set_properties(prop2index, prop2value)
+        self.set_properties(prop2index=prop2index, prop2value=prop2value)
+
+        # TODO check this works on all Canon models.
+        self.model = self.get_property('d402')
+
         self._connected = True
 
-    def take_observation(self, observation, headers=None, filename=None, *args, **kwargs):
-        """Take an observation
+    def _start_exposure(self,
+                        seconds=None,
+                        filename=None,
+                        dark=False,
+                        header=None,
+                        iso=100,
+                        *args, **kwargs):
+        """Start the exposure.
 
-        Gathers various header information, sets the file path, and calls
-        `take_exposure`. Also creates a `threading.Event` object and a
-        `threading.Timer` object. The timer calls `process_exposure` after the
-        set amount of time is expired (`observation.exptime + self.readout_time`).
-
-        Note:
-            If a `filename` is passed in it can either be a full path that includes
-            the extension, or the basename of the file, in which case the directory
-            path and extension will be added to the `filename` for output
-
-        Args:
-            observation (~pocs.scheduler.observation.Observation): Object
-                describing the observation
-            headers (dict): Header data to be saved along with the file
-            filename (str, optional): Filename for saving, defaults to ISOT time stamp
-            **kwargs (dict): Optional keyword arguments (`exptime`)
-
-        Returns:
-            threading.Event: An event to be set when the image is done processing
-        """
-        # To be used for marking when exposure is complete (see `process_exposure`)
-        observation_event = Event()
-
-        exptime, file_path, image_id, metadata = self._setup_observation(observation,
-                                                                         headers,
-                                                                         filename,
-                                                                         **kwargs)
-
-        exposure_event = self.take_exposure(seconds=exptime, filename=file_path)
-
-        # Add most recent exposure to list
-        if self.is_primary:
-            if 'POINTING' in headers:
-                observation.pointing_images[image_id] = file_path.replace('.cr2', '.fits')
-            else:
-                observation.exposure_list[image_id] = file_path.replace('.cr2', '.fits')
-
-        # Process the image after a set amount of time
-        wait_time = exptime + self.readout_time
-
-        t = Timer(wait_time, self.process_exposure, (metadata, observation_event))
-        t.name = f'{self.name}Thread'
-        t.start()
-
-        return observation_event
-
-    def _start_exposure(self, seconds=None, filename=None, dark=None, header=None, *args, **kwargs):
-        """Take an exposure for given number of seconds and saves to provided filename
-
-        Note:
-            See `scripts/take-pic.sh`
-
-            Tested With:
-                * Canon EOS 100D
+        Tested With:
+            * Canon EOS 100D
 
         Args:
-            seconds (u.second, optional): Length of exposure
-            filename (str, optional): Image is saved to this filename
+            seconds (u.second, optional): Length of exposure.
+            filename (str, optional): Image is saved to this filename.
+            header (dict, optional): The metadata to be added as FITS headers.
+            iso (int, optional): The ISO setting to use for the exposure, default 100.
         """
-        script_path = os.path.expandvars('$POCS/scripts/take-pic.sh')
-
         # Make sure we have just the value, no units
         seconds = get_quantity_value(seconds)
 
-        run_cmd = [script_path, self.port, str(seconds), filename]
+        cmd_args = [
+            f'--set-config-index', 'shutterspeed=0',
+            f'--set-config', f'iso={iso}',
+            f'--set-config-index', 'eosremoterelease=2',
+            f'--wait-event={int(seconds):d}s',
+            f'--set-config-index', 'eosremoterelease=4',
+            f'--wait-event-and-download=1s',
+            f'--filename', f'{filename}'
+        ]
 
-        # Take Picture
         try:
-            self._is_exposing_event.set()
-            self._exposure_proc = subprocess.Popen(run_cmd,
-                                                   stdout=subprocess.PIPE,
-                                                   stderr=subprocess.PIPE,
-                                                   universal_newlines=True)
+            self.command(cmd_args)
         except error.InvalidCommand as e:
             self.logger.warning(e)
-        finally:
+        else:
             readout_args = (filename, header)
             return readout_args
-
-    def _readout(self, cr2_path=None, info=None):
-        """Reads out the image as a CR2 and converts to FITS"""
-        self.logger.debug(f"Converting CR2 -> FITS: {cr2_path}")
-        fits_path = cr2_utils.cr2_to_fits(cr2_path, headers=info, remove_cr2=False)
-        return fits_path
-
-    def _process_fits(self, file_path, info):
-        """
-        Add FITS headers from info the same as images.cr2_to_fits()
-        """
-        file_path = file_path.replace('.cr2', '.fits')
-        return super()._process_fits(file_path, info)
-
-    def _poll_exposure(self, readout_args):
-        timer = CountdownTimer(duration=self._timeout)
-        try:
-            try:
-                # See if the command has finished.
-                while self._exposure_proc.poll() is None:
-                    # Sleep if not done yet.
-                    timer.sleep()
-            except subprocess.TimeoutExpired:
-                self.logger.warning(f'Timeout on exposure process for {self.name}')
-                self._exposure_proc.kill()
-                outs, errs = self._exposure_proc.communicate(timeout=10)
-                if errs is not None and errs > '':
-                    self.logger.error(f'Camera exposure errors: {errs}')
-        except (RuntimeError, error.PanError) as err:
-            # Error returned by driver at some point while polling
-            self.logger.error('Error while waiting for exposure on {}: {}'.format(self, err))
-            raise err
-        else:
-            # Camera type specific readout function
-            self._readout(*readout_args)
-        finally:
-            self.logger.debug(f'Setting exposure event for {self.name}')
-            self._exposure_proc = None
-            self._is_exposing_event.clear()  # Make sure this gets set regardless of readout errors
