@@ -9,7 +9,6 @@ import astropy.units as u
 from panoptes.pocs.camera.simulator.dslr import Camera as SimCamera
 from panoptes.pocs.camera.sdk import AbstractSDKDriver, AbstractSDKCamera
 from panoptes.utils.config.client import get_config
-from panoptes.utils.logging import logger
 
 
 class SDKDriver(AbstractSDKDriver):
@@ -21,14 +20,14 @@ class SDKDriver(AbstractSDKDriver):
         return "Simulated SDK Driver v0.001"
 
     def get_devices(self):
-        logger.debug(f'Getting camera device connection config for {self}')
+        self.logger.debug(f'Getting camera device connection config for {self}')
         camera_devices = dict()
         for cam_info in get_config('cameras.devices'):
             name = cam_info.get('name') or cam_info.get('model')
             port = cam_info.get('port') or cam_info.get('serial_number')
             camera_devices[name] = port
 
-        logger.trace(f'camera_devices={camera_devices!r}')
+        self.logger.trace(f'camera_devices={camera_devices!r}')
 
         return camera_devices
 
