@@ -1,13 +1,12 @@
 from contextlib import suppress
 from glob import glob
 
-from panoptes.pocs.mount.mount import AbstractMount  # pragma: no flakes
+from panoptes.pocs.mount.mount import AbstractMount  # noqa
 from panoptes.pocs.utils.location import create_location_from_config
 from panoptes.pocs.utils.logger import get_logger
 from panoptes.utils import error
 from panoptes.utils.library import load_module
-from panoptes.utils.config.client import get_config
-from panoptes.utils.config.client import set_config
+from panoptes.utils.config.client import get_config, set_config
 
 logger = get_logger()
 
@@ -57,19 +56,20 @@ def create_mount_from_config(mount_info=None,
         site_details = create_location_from_config()
         earth_location = site_details['earth_location']
 
+    brand = mount_info.get('brand')
     driver = mount_info.get('driver')
+    model = mount_info.get('model', driver)
     if not driver or not isinstance(driver, str):
         raise error.MountNotFound('Mount info in config is missing a driver name.')
 
-    model = mount_info.get('model', driver)
-    logger.debug(f'Mount: driver={driver} model={model}')
+    logger.debug(f'Mount: {brand=} {driver=} {model=}')
 
     # Check if we should be using a simulator
     use_simulator = 'mount' in get_config('simulator', default=[])
     logger.debug(f'Mount is simulator: {use_simulator}')
 
     # Create simulator if requested
-    if use_simulator or (driver == 'simulator'):
+    if use_simulator or ('simulator' in driver):
         logger.debug(f'Creating mount simulator')
         return create_mount_simulator(mount_info=mount_info, earth_location=earth_location)
 
@@ -78,8 +78,7 @@ def create_mount_from_config(mount_info=None,
         port = mount_info['serial']['port']
         logger.info(f'Looking for {driver} on {port}.')
         if port is None or len(glob(port)) == 0:
-            msg = f'Mount port ({port}) not available. Use simulator = mount for simulator.'
-            raise error.MountNotFound(msg=msg)
+            raise error.MountNotFound(msg=f'Mount {port=} not available.')
     except KeyError:
         # See Issue 866
         if model == 'bisque':
@@ -88,9 +87,9 @@ def create_mount_from_config(mount_info=None,
             msg = 'Mount port not specified in config file. Use simulator=mount for simulator.'
             raise error.MountNotFound(msg=msg)
 
-    logger.debug(f'Loading mount driver: pocs.mount.{driver}')
+    logger.debug(f'Loading mount {driver=}')
     try:
-        module = load_module(f'panoptes.pocs.mount.{driver}')
+        module = load_module(driver)
     except error.NotFound as e:
         raise error.MountNotFound(e)
 
@@ -104,6 +103,7 @@ def create_mount_from_config(mount_info=None,
 
 def create_mount_simulator(mount_info=None,
                            earth_location=None,
+                           db_type='memory',
                            *args, **kwargs):
     # Remove mount simulator
     current_simulators = get_config('simulator', default=[])
@@ -113,7 +113,7 @@ def create_mount_simulator(mount_info=None,
 
     mount_config = mount_info or {
         'model': 'Mount Simulator',
-        'driver': 'simulator',
+        'driver': 'panoptes.pocs.mount.simulator',
         'serial': {
             'port': '/dev/FAKE'
         }
@@ -124,14 +124,14 @@ def create_mount_simulator(mount_info=None,
 
     earth_location = earth_location or create_location_from_config()['earth_location']
 
-    logger.debug(f"Loading mount driver: pocs.mount.{mount_config['driver']}")
+    logger.debug(f"Loading mount driver: {mount_config['driver']}")
     try:
-        module = load_module(f"panoptes.pocs.mount.{mount_config['driver']}")
+        module = load_module(f"{mount_config['driver']}")
     except error.NotFound as e:
         raise error.MountNotFound(f'Error loading mount module: {e!r}')
 
-    mount = module.Mount(earth_location, *args, **kwargs)
+    mount_obj = module.Mount(earth_location, db_type=db_type, *args, **kwargs)
 
     logger.success(f"{mount_config['driver'].title()} mount created")
 
-    return mount
+    return mount_obj
