@@ -1,8 +1,11 @@
-
 def on_enter(event_data):
     """ The unit is tracking the target. Proceed to observations. """
     pocs = event_data.model
     pocs.next_state = 'parking'
+
+    if pocs.get_config('mount.settings.update_tracking', False):
+        pocs.next_state = 'observing'
+        return
 
     # If we came from pointing then don't try to adjust
     if event_data.transition.source != 'pointing':
